PIDE: Handle the Perspective correctly.
A perspective message is sent whenever the user changes what s/he is
looking at: this allows the STM (in certain cases) to re-prioritize the
tasks belonging to this perspective. In particular, jEdit remembers the
user's position in the buffer when it quits, and it will therefore send
a mid-buffer perspective whenever the editor is reopened: the user can
start working where s/he left off almost immediately.