PIDE: Query overlays now work as expected.
In the previous commit(s), the overlay was not added to the common
prefix of the previous document version and the new version. This meant
that queries where not executed on the document that was already
available (the common case).

To achieve this, the update and execute functions have been restructured to directly
build a task queue, which is then passed on to pide_top's task queue.