Revert "Place user-facing Pp functions in their own module."
While it might still be a better way to do feedback, it was not enough
to break all the cyclic dependencies that introducing a "vernac_entries"
based message in the Feedback module caused (it required serializing a
Names.Id.t, which is part of the kernel, and the kernel requires the
clib.mllib, which includes Feedback, which needs to serialize the
Names.Id.t, which closes the loop).

Going forwards, I am encoding the Vernac_entries payload into XML the
moment STM wants to emit it, and maybe reconstruct the payload when PIDE
receives the message (it might be sufficient to do the transformation on
the XML level, because we are just translating to the client language in
any case).

This reverts commit 9425c65d6b086841ab33d9e91f1ce03cdd3a661b.