PIDE: Moved printers from protocol to document.
This is more appropriate, because prints can refer to the document state for
more information. For example, the lookup map for references can become
part of the document state.

The concrete occasion for this refactoring is that I need to print query
results using an instance id, if available, which can be known inside
the document state (it comes with an update), but which I do not want
to export as some mutable state.