PIDE: Deal with new version of define_command
This PIDE protocol function now has a Coq-specific implementation, which
reports whether a command span should be ignored (as reported by the
parser). The concrete use case is that when a script ends in a comment,
this comment is sent as a command by PIDE, but it should not be
processed by Coq, as it leads into an anomaly.