Factored Error_printer to printer spec.
This also adds a make_pos function to the printer spec, as some printers
give more localized information.

Note that the spec might actually introduce a bit more boilerplate, but
it does allow us uniformly handling things like query results: it might
be a bit more syntactically heavy, but the behaviour is a lot more
uniform and easy to change.