Factored overlay into mandatory and query parts.
This factoring allows us to set some print functions to always execute
(such as print goal state/error) and others to only execute when a query
is specified by the editor.

During the factoring, I also took the opportunity to remove an argument
to set_overlay that was no longer needed.