Schedule a "FeedbackGoals" query on each state.
The actual query still needs to be implemented, so the code gives errors
on each command, currently.

There is a TODO here for the STM code: it would be nice to be able to
execute arbitrary (unit-codomained) functions on a state, instead of
having to provide a syntax for functions that are only used on the OCaml
side.