do not erase goal prints when adding queries via the overlay
The scala layer holds a map

  span_id |-> exec_ids

to any exec_id some feedback can be attached.  When one updates the list
if exec_ids, old ids no more present gets their associated feedback removed.

With this patch once the goal are printed the exec id for the corresponding
query is never removed, hence the data is not lost.